Move Database Cog from Utils to main cogs file for clarity. - Removed terminal logging methods from DatabaseCog and added it to UtilityCog - Remove show_all_commands command from OguriCapCog and integrate it into UtilityCog. - Tempfix for manual parsing in add_scores and overwrite_scores methods
