+++ cogs/tests/test_oguri_cap.py
@@ -75,17 +75,3 @@ async def test_snack_command_sends_embed(cog):
         "I can't go all-out on an empty stomach.",
         "I run. I earned this."
     ]
-
-@pytest.mark.asyncio
-async def test_celebrate_victory_command_sends_embed(cog):
-    """Test the celebrate_victory command sends an embed with a victory gif."""
-    ctx = mock.AsyncMock()
-    with mock.patch('random.choice', return_value=cog.gifs['victory'][0]):
-        await cog.celebrate_victory.callback(cog, ctx)
-    # Check that ctx.send was called with an embed
-    ctx.send.assert_called_once()
-    call_args = ctx.send.call_args
-    embed = call_args[1]['embed'] if call_args[1] else call_args[0][0]
-    
-    assert embed.title == "A fine victory for Mr. Tricks"
-    assert embed.description == "AND THE GOAT WINS AGAIN 🎉"
